chore(ProductService): Изменение ProductModel, DataBase, миграции
ProductModel:

* Изменен тип поля Rating с decimal -> double
* Переименовано имя свойства SellerId -> ShopId
* Переименовано на ShopId везде где был SellerId

DataBase:

* Изменения Constraint - Check title length для таблицы categories, products
* Переименована колонка таблицы products seller_id -> shop_id
* Добавлены миграции, автоматическое применение новых миграций в Programm.cs

+++ services/ProductService/src/Domain/Product/ProductConstraints.cs
@@ -8,10 +8,10 @@
         public const int MIN_Description_LENGTH = 0;
         public const int MAX_Description_LENGTH = 6000;
 
-        public const decimal MIN_PRICE = 50;
-        public const decimal MAX_PRICE = 10000000;
+        public const decimal MIN_PRICE = 50M;
+        public const decimal MAX_PRICE = 10000000M;
 
-        public const decimal MIN_RATING = 0.0M;
-        public const decimal MAX_RATING = 5.0M;
+        public const double MIN_RATING = 0.0;
+        public const double MAX_RATING = 5.0;
     }
 }
